third_party/blink/web_tests/virtual/single-import-map/external/wpt/import-maps/multiple-import-maps/resolution-consistency-in-module-tree-inline-expected.txt

This is a testharness.js-based test.
[FAIL] Module tree that started to download before a new import map should still take it into account
  assert_array_equals: Import should use the new import map expected property 0 to be "log:B" but got "log:A" (expected array ["log:B"] got ["log:A"])
Harness: the test ran to completion.
